Adds a dummy (not to be used) simple auth server
Identifying to this auth server, you can get a token, that can be later
used on the socket connection

Right now, the token is just the md5 of the device id. No proper auth is
being done, so this implementation is just for testing purposes. Now
that this can be tested, we can properly discuss auth schemes, possible
problems with this authentication workflow, and blah blah.

The auth server is an express instance (HTTP server)